fix: restore auth middleware in prod builds and require nextauth_secret (#1400)
- rename proxy.ts back to middleware.ts: next 16.2 webpack prod builds
  silently exclude proxy.ts (empty middleware-manifest.json), so deployed
  builds shipped without the auth gate
- set nextauthoptions.secret explicitly and fail fast when nextauth_secret
  is missing at runtime — withauth has no dev fallback secret, so a missing
  env var means an infinite login loop
- add local-dev nextauth_secret to the tracked local env template
- app header logo links to /atlases (full header is authenticated-only);
  landing header re-points it to /
