+++ site-config/hca-atlas-tracker/local/.env
@@ -2,3 +2,7 @@ NEXT_PUBLIC_SITE_CONFIG='hca-atlas-tracker-local'
 SERVICE_ACCOUNT_EMAIL='tracker-dev@hca-atlas-tracker-dev.iam.gserviceaccount.com'
 HCA_VALIDATION_TOOLS_URL='https://2w532lsuzmewkdx7d4eutr2ly40vxciu.lambda-url.us-east-1.on.aws/'
 AWS_DATA_BUCKET='hca-atlas-tracker-data-dev'
+# Shared JWT secret for NextAuth: the API route, getServerSession, and the
+# auth middleware must all use the same value. Local-dev-only value — deployed
+# environments set their own NEXTAUTH_SECRET in the task environment.
+NEXTAUTH_SECRET='local-dev-only-nextauth-secret-do-not-use-deployed'
