+++ site-config/hca-atlas-tracker/local/authentication/next-auth-config.ts
@@ -28,6 +28,16 @@ if (!IS_BUILD_PHASE) {
       "GOOGLE_AUTH environment variable must contain a JSON object with non-empty `client_id` and `client_secret`.",
     );
   }
+  // The NextAuth API route derives a fallback secret in dev when none is
+  // configured, but `withAuth` in `middleware.ts` only reads
+  // NEXTAUTH_SECRET — with no shared secret the middleware rejects every
+  // session and login loops back to the sign-in page forever. Fail fast
+  // instead of shipping that silent loop.
+  if (!process.env.NEXTAUTH_SECRET) {
+    throw new Error(
+      "NEXTAUTH_SECRET environment variable must be set; the auth middleware cannot validate sessions without it.",
+    );
+  }
 }
 
 export const nextAuthOptions: NextAuthOptions = {
@@ -41,6 +51,10 @@ export const nextAuthOptions: NextAuthOptions = {
       clientSecret: googleAuthParams.client_secret ?? "",
     }),
   ],
+  // Explicit so the API route, `getServerSession` callers, and the
+  // middleware all validate sessions with the same secret (the API route
+  // would otherwise silently derive its own fallback in dev).
+  secret: process.env.NEXTAUTH_SECRET,
   session: {
     maxAge: SESSION_MAX_AGE / 1000,
   },
